backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
sqlalchemy>=2.0.28
psycopg[binary]
asyncpg>=0.29.0
python-dotenv>=1.0.1
httpx>=0.27.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
feedparser==6.0.10
beautifulsoup4==4.12.3
google-generativeai==0.4.1
tenacity==8.2.3
apscheduler==3.10.4
python-multipart==0.0.9
email-validator==2.1.1
jinja2==3.1.3
openai>=1.0.0
python-jose[cryptography]
resend
